+++ src/lib-sasl/dsasl-client.h
@@ -23,14 +23,12 @@ void dsasl_client_free(struct dsasl_client **client);
 
 /* Call for server input. */
 int dsasl_client_input(struct dsasl_client *client,
-		       const unsigned char *input,
-		       unsigned int input_len,
+		       const unsigned char *input, size_t input_len,
 		       const char **error_r);
 /* Call for getting server output. Also used to get the initial SASL response
    if supported by the protocol. */
 int dsasl_client_output(struct dsasl_client *client,
-			const unsigned char **output_r,
-			unsigned int *output_len_r,
+			const unsigned char **output_r, size_t *output_len_r,
 			const char **error_r);
 
 void dsasl_clients_init(void);
