+++ src/lib-sasl/mech-plain.c
@@ -11,7 +11,7 @@ struct plain_dsasl_client {
 
 static int
 mech_plain_input(struct dsasl_client *_client,
-		 const unsigned char *input ATTR_UNUSED, unsigned int input_len,
+		 const unsigned char *input ATTR_UNUSED, size_t input_len,
 		 const char **error_r)
 {
 	struct plain_dsasl_client *client =
@@ -31,7 +31,7 @@ mech_plain_input(struct dsasl_client *_client,
 
 static int
 mech_plain_output(struct dsasl_client *_client,
-		  const unsigned char **output_r, unsigned int *output_len_r,
+		  const unsigned char **output_r, size_t *output_len_r,
 		  const char **error_r)
 {
 	struct plain_dsasl_client *client =
